revise select component structure
